requirements.txt
qq-botpy>=2.0.0
pyyaml>=6.0
asyncio>=3.4.3
requests>=2.28.0 